initial code clean-up, move tests to other files (not tested yet)
